Make ssh-copy-id selinux-aware
When ssh-copy-id is used to transfer keys to SELinux-enabled machine,
'~/.ssh' and '~/.ssh/authorized_keys' may be created with improper
SELinux contexts.  This patch makes ssh-copy-id try to restore the
right context in case 'restorecon' binary is available.  I also make
all the commands instead the restorecon fail immediately, even though
such cases should be very rare (permission or other problems when
creating '~/.ssh' or '~/.ssh/authorized_keys').  That makes
ssh-copy-id output only one (appropriate) error message (for example
when mkdir fails).

I hope the ':' (a.k.a. 'true') at the end is POSIX-compatible, but
'true' can be there too in case ':' is not supported somewhere.  The
purpose of that command is not to fail in case there is no
'restorecon' available.  Otherwise missing 'restorecon' would mean
that the script ends right after the ssh connection is over.

To make it more readable, I re-indented the code a bit as well.