Keep the default no-options behavior
When no options were specified on a machine without ssh-agent,
ssh-copy-id reported "no keys found" error.  This patch makes the
script behave as old versions of ssh-copy-id from upstream openssh
where the default key(s) were used even when no '-i' was specified.

I'm not sure this is wanted behavior from others point of view, but it
makes sense to me not to break old scripts using upstream's openssh's
ssh-copy-id (in case there are any :-) ) when it takes so little.