add message to allow one to distinguish the reasons one might be prompted
without this, one sees a message indicating that some login tests
are going to happen, then might get prompted immediately if there's
no ssh-agent running, or might not.

This INFO message will then appear making sure that one knows that if
prompted it is to allow the key to be installed -- without that it's
possible to assume that we're still in the key checking phase, and
simply hit return in order to save time and force the key to be
installed (which it won't be, because you just sabotaged the attempt).